Add KFontLoader Support texture array load/save

+++ KlayGE/Core/Include/KlayGE/RenderFactory.hpp
@@ -47,13 +47,13 @@ namespace KlayGE
 
 		RenderEngine& RenderEngineInstance();
 
-		virtual TexturePtr MakeTexture1D(uint32_t width, uint16_t numMipMaps, uint16_t array_size,
+		virtual TexturePtr MakeTexture1D(uint32_t width, uint32_t numMipMaps, uint32_t array_size,
 			ElementFormat format, uint32_t sample_count, uint32_t sample_quality, uint32_t access_hint, ElementInitData* init_data) = 0;
-		virtual TexturePtr MakeTexture2D(uint32_t width, uint32_t height, uint16_t numMipMaps, uint16_t array_size,
+		virtual TexturePtr MakeTexture2D(uint32_t width, uint32_t height, uint32_t numMipMaps, uint32_t array_size,
 			ElementFormat format, uint32_t sample_count, uint32_t sample_quality, uint32_t access_hint, ElementInitData* init_data) = 0;
-		virtual TexturePtr MakeTexture3D(uint32_t width, uint32_t height, uint32_t depth, uint16_t numMipMaps, uint16_t array_size,
+		virtual TexturePtr MakeTexture3D(uint32_t width, uint32_t height, uint32_t depth, uint32_t numMipMaps, uint32_t array_size,
 			ElementFormat format, uint32_t sample_count, uint32_t sample_quality, uint32_t access_hint, ElementInitData* init_data) = 0;
-		virtual TexturePtr MakeTextureCube(uint32_t size, uint16_t numMipMaps, uint16_t array_size,
+		virtual TexturePtr MakeTextureCube(uint32_t size, uint32_t numMipMaps, uint32_t array_size,
 			ElementFormat format, uint32_t sample_count, uint32_t sample_quality, uint32_t access_hint, ElementInitData* init_data) = 0;
 		virtual FrameBufferPtr MakeFrameBuffer() = 0;
