+++ KlayGE/Plugins/Include/KlayGE/D3D9/D3D9Texture.hpp
@@ -103,7 +103,7 @@ namespace KlayGE
 	class D3D9Texture1D : public D3D9Texture
 	{
 	public:
-		D3D9Texture1D(uint32_t width, uint16_t numMipMaps, uint16_t array_size,
+		D3D9Texture1D(uint32_t width, uint32_t numMipMaps, uint32_t array_size,
 			ElementFormat format, uint32_t sample_count, uint32_t sample_quality, uint32_t access_hint, ElementInitData* init_data);
 
 		uint32_t Width(int level) const;
@@ -144,7 +144,7 @@ namespace KlayGE
 	class D3D9Texture2D : public D3D9Texture
 	{
 	public:
-		D3D9Texture2D(uint32_t width, uint32_t height, uint16_t numMipMaps, uint16_t array_size,
+		D3D9Texture2D(uint32_t width, uint32_t height, uint32_t numMipMaps, uint32_t array_size,
 			ElementFormat format, uint32_t sample_count, uint32_t sample_quality, uint32_t access_hint, ElementInitData* init_data);
 
 		uint32_t Width(int level) const;
@@ -188,7 +188,7 @@ namespace KlayGE
 	class D3D9Texture3D : public D3D9Texture
 	{
 	public:
-		D3D9Texture3D(uint32_t width, uint32_t height, uint32_t depth, uint16_t numMipMaps, uint16_t array_size,
+		D3D9Texture3D(uint32_t width, uint32_t height, uint32_t depth, uint32_t numMipMaps, uint32_t array_size,
 			ElementFormat format, uint32_t sample_count, uint32_t sample_quality, uint32_t access_hint, ElementInitData* init_data);
 
 		uint32_t Width(int level) const;
@@ -236,7 +236,7 @@ namespace KlayGE
 	class D3D9TextureCube : public D3D9Texture
 	{
 	public:
-		D3D9TextureCube(uint32_t size, uint16_t numMipMaps, uint16_t array_size,
+		D3D9TextureCube(uint32_t size, uint32_t numMipMaps, uint32_t array_size,
 			ElementFormat format, uint32_t sample_count, uint32_t sample_quality, uint32_t access_hint, ElementInitData* init_data);
 
 		uint32_t Width(int level) const;
