+++ KlayGE/Plugins/Src/Render/D3D10/D3D10Texture3D.cpp
@@ -42,7 +42,7 @@
 
 namespace KlayGE
 {
-	D3D10Texture3D::D3D10Texture3D(uint32_t width, uint32_t height, uint32_t depth, uint16_t numMipMaps, uint16_t array_size, ElementFormat format,
+	D3D10Texture3D::D3D10Texture3D(uint32_t width, uint32_t height, uint32_t depth, uint32_t numMipMaps, uint32_t array_size, ElementFormat format,
 						uint32_t sample_count, uint32_t sample_quality, uint32_t access_hint, ElementInitData* init_data)
 					: D3D10Texture(TT_3D, sample_count, sample_quality, access_hint)
 	{
@@ -68,7 +68,7 @@ namespace KlayGE
 		std::vector<D3D10_SUBRESOURCE_DATA> subres_data(numMipMaps_);
 		if (init_data != NULL)
 		{
-			for (int i = 0; i < numMipMaps_; ++ i)
+			for (uint32_t i = 0; i < numMipMaps_; ++ i)
 			{
 				subres_data[i].pSysMem = init_data[i].data;
 				subres_data[i].SysMemPitch = init_data[i].row_pitch;
@@ -268,7 +268,7 @@ namespace KlayGE
 	{
 		d3dTexture3D_->GetDesc(&desc_);
 
-		numMipMaps_ = static_cast<uint16_t>(desc_.MipLevels);
+		numMipMaps_ = desc_.MipLevels;
 		BOOST_ASSERT(numMipMaps_ != 0);
 
 		widthes_.resize(numMipMaps_);
