+++ KlayGE/Plugins/Src/Render/D3D11/D3D11TextureCube.cpp
@@ -42,7 +42,7 @@
 
 namespace KlayGE
 {
-	D3D11TextureCube::D3D11TextureCube(uint32_t size, uint16_t numMipMaps, uint16_t array_size, ElementFormat format,
+	D3D11TextureCube::D3D11TextureCube(uint32_t size, uint32_t numMipMaps, uint32_t array_size, ElementFormat format,
 						uint32_t sample_count, uint32_t sample_quality, uint32_t access_hint, ElementInitData* init_data)
 					: D3D11Texture(TT_Cube, sample_count, sample_quality, access_hint)
 	{
@@ -69,7 +69,7 @@ namespace KlayGE
 		{
 			for (int face = 0; face < 6; ++ face)
 			{
-				for (int i = 0; i < numMipMaps_; ++ i)
+				for (uint32_t i = 0; i < numMipMaps_; ++ i)
 				{
 					subres_data[face * numMipMaps_ + i].pSysMem = init_data[face * numMipMaps_ + i].data;
 					subres_data[face * numMipMaps_ + i].SysMemPitch = init_data[face * numMipMaps_ + i].row_pitch;
@@ -267,13 +267,13 @@ namespace KlayGE
 	{
 		d3dTextureCube_->GetDesc(&desc_);
 
-		numMipMaps_ = static_cast<uint16_t>(desc_.MipLevels);
-		array_size_ = static_cast<uint16_t>(desc_.ArraySize / 6);
+		numMipMaps_ = desc_.MipLevels;
+		array_size_ = desc_.ArraySize / 6;
 		BOOST_ASSERT(numMipMaps_ != 0);
 
 		widthes_.resize(numMipMaps_);
 		widthes_[0] = desc_.Width;
-		for (uint16_t level = 1; level < numMipMaps_; ++ level)
+		for (uint32_t level = 1; level < numMipMaps_; ++ level)
 		{
 			widthes_[level] = widthes_[level - 1] / 2;
 		}
