+++ KlayGE/Plugins/Src/Render/OpenGL/OGLTexture.cpp
@@ -42,7 +42,7 @@
 
 namespace KlayGE
 {
-	OGLTexture::OGLTexture(TextureType type, uint16_t array_size, uint32_t sample_count, uint32_t sample_quality, uint32_t access_hint)
+	OGLTexture::OGLTexture(TextureType type, uint32_t array_size, uint32_t sample_count, uint32_t sample_quality, uint32_t access_hint)
 					: Texture(type, sample_count, sample_quality, access_hint)
 	{
 		if ((array_size > 1) && (!(glloader_GL_VERSION_3_0() || glloader_GL_EXT_texture_array())))
